Add state visibility between Executions and Operations.
Added a state variable in Operation, which tracks the state of the
Execution currently running. In the Execution implementations, before
starting, the state for all the operations in the scenario is set
to BEGINNING. Upon completion (if an error) or just prior to shutting
down the Execution, the state for all the Operations is changed to
ENDING.

This allows for user implemented operations to catch RuntimeExceptions
(which may be caused by thread interrupts, etc) and discard them if
the Operation is already shutting down.

Additionally, a small change was made in HtmlReporter. The Histogram
was being arbitrarily corrected by 1000000L; in certain cases this
cases an IndexOutOfBounds exception, or other weird exceptions. In this
case, the historgram is generated uncorrected. Anyone with a better idea
is welcome to contribute.