Add per thread Histogram's to facilitate multi-thread testing.
Previously, a single Histogram instance was updated by all active
working threads. Now, a RainfalHistogramSink class sits in front,
and allocates per thread Histograms as needed for the recording
of values.

WHen it comes time to report, we make use of the Histogram summing
capabilities to create a single histogram object from all the
per thread instances.

This allows for both fast and concurrent updating of the stats.